backend/requirements.txt
fastapi==0.104.1
pydantic[email]
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary
argon2-cffi